Changed the order of the things in the Summit Guide
A user had suggested that the Summit Guide was hard to navigate
because the important information was all over the place, the
unimportant information is close to the top, and overall is hard
to navigate for the new user trying to get familiar with Summit
through the Summit guide. This commit moves down the large amounts
of info on the GPUs and Burst Buffers to the bottom, moves up the
info about jsrun higher in the Running Jobs section. These changes
are intended to improve the flow of the Guide so that there is a logical
sequence of the information being presented, and that the relevant
information is presented at the right points.